Move verification code to a separate file
Remove a failed verification attempt in verifier.py at the same time.
